+++ packages/select/src/helperComponents/TagSelect/Option/Option.tsx
@@ -1,22 +1,20 @@
 import { cx } from '@linaria/core';
-import isEqual from 'lodash.isequal';
-import { useEffect, useRef, useState } from 'react';
-import { components as ReactSelectComponents } from 'react-select';
-
 import { CircleCheckSVG, DeleteSVG, EditSVG } from '@sbercloud/icons';
 import { Button } from '@sbercloud/uikit-react-button';
-import { Tag } from '@sbercloud/uikit-react-tag';
+import isEqual from 'lodash.isequal';
+import { useEffect, useMemo, useRef, useState } from 'react';
+import { components as ReactSelectComponents } from 'react-select';
 
 import { ColorPicker, OptionTypeColor } from '../../../components';
+import { TagName } from '../TagName';
 import {
-  StyledTag,
+  NotValidMessage,
   StyledTagButton,
   StyledTagButtonWrapper,
   StyledTagOption,
   StyledTagOptionLabel,
   colorPickerClassName,
   optionClass,
-  tagInputClassName,
 } from './styled';
 
 export const Option = (props: React.ComponentProps<typeof ReactSelectComponents.Option>): JSX.Element => {
@@ -24,12 +22,26 @@ export const Option = (props: React.ComponentProps<typeof ReactSelectComponents.
     data,
     innerRef,
     className,
-    selectProps: { onTagChange, approveDeleting, isSelected, colorDropdownPlacement, setMenuListBlockScroll },
+    selectProps: {
+      onTagChange,
+      approveDeleting,
+      isSelected,
+      colorDropdownPlacement,
+      setMenuListBlockScroll,
+      editableTagName,
+      validator,
+      validateMessage,
+    },
   } = props;
 
   const [isEdit, setEdit] = useState(false);
   const [tag, setTag] = useState(data);
   const [colorVal, setColorVal] = useState({ value: tag.color });
+
+  const notValid = useMemo(
+    () => editableTagName && validator && !validator(tag.label),
+    [validator, tag, editableTagName],
+  );
 
   if (isEdit) {
     props.innerProps.onClick = (e): void => {
@@ -71,15 +83,10 @@ export const Option = (props: React.ComponentProps<typeof ReactSelectComponents.
       <StyledTagOption data-is-edit={isEdit} ref={innerRef} onBlur={handleBlur}>
         {isEdit ? (
           <>
-            <StyledTag
-              inputRef={(ref): void => {
-                ref?.focus();
-              }}
-              color={tag.color}
-              value={tag.label}
-              type={Tag.types.Input}
-              inputClassNames={tagInputClassName}
-              onChange={(e: { target: { value: any } }): void => {
+            <TagName
+              tag={tag}
+              editableTagName={editableTagName}
+              onChange={(e: { target: { value: string } }) => {
                 setTag({ ...tag, label: e.target.value });
               }}
             />
@@ -127,7 +134,9 @@ export const Option = (props: React.ComponentProps<typeof ReactSelectComponents.
               <StyledTagButton
                 size={Button.sizes.xs}
                 variant={Button.variants.Transparent}
+                data-disabled={notValid || undefined}
                 onClick={(): void => {
+                  if (notValid) return;
                   setEdit(false);
                   if (!isEqual(data, tag) && tag.label) {
                     const dataIndex = props.options.indexOf(data);
@@ -163,6 +172,7 @@ export const Option = (props: React.ComponentProps<typeof ReactSelectComponents.
           </>
         )}
       </StyledTagOption>
+      {notValid && isEdit && <NotValidMessage>{validateMessage}</NotValidMessage>}
     </ReactSelectComponents.Option>
   );
 };