+++ scripts/utils/filesTemplate.js
@@ -22,6 +22,7 @@ const packageJson = ({ user, email, packageTitle, packageName, packageDescriptio
     name: `@sbercloud/uikit-${packageName}`,
     title: `${packageTitle}`,
     version: '0.0.0',
+    sideEffects: false,
     description: `${packageDescription}`,
     main: './dist/esm/index.js',
     module: './dist/esm/index.js',