+++ packages/checkbox/CHANGELOG.md
@@ -3,6 +3,25 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.1.15](https://bitbucket.sbercloud.tech/projects/SBERCLOUD_UI/repos/uikit2.0/compare/diff?targetBranch=refs/tags/@sbercloud/uikit-react-checkbox@0.1.14&sourceBranch=refs/tags/@sbercloud/uikit-react-checkbox@0.1.15) (2021-06-25)
+
+**Note:** Version bump only for package @sbercloud/uikit-react-checkbox
+
+
+
+
+
+## [0.1.14](https://bitbucket.sbercloud.tech/projects/SBERCLOUD_UI/repos/uikit2.0/compare/diff?targetBranch=refs/tags/@sbercloud/uikit-react-checkbox@0.1.13&sourceBranch=refs/tags/@sbercloud/uikit-react-checkbox@0.1.14) (2021-06-25)
+
+
+### Bug Fixes
+
+* sideEffects flag in package.json was set for correct tree shaking ([8d72148](https://bitbucket.sbercloud.tech/projects/SBERCLOUD_UI/repos/uikit2.0/commits/8d721482e140bb8ea41d62d7bc90769c642b7f0f))
+
+
+
+
+
 ## [0.1.13](https://bitbucket.sbercloud.tech/projects/SBERCLOUD_UI/repos/uikit2.0/compare/diff?targetBranch=refs/tags/@sbercloud/uikit-react-checkbox@0.1.12&sourceBranch=refs/tags/@sbercloud/uikit-react-checkbox@0.1.13) (2021-06-24)
 
 **Note:** Version bump only for package @sbercloud/uikit-react-checkbox